+++ pkg/analysis_server/test/domain_analysis_test.dart
@@ -99,9 +99,6 @@ class AnalysisDomainPubTest extends _AnalysisDomainTest {
     // Both a.dart and b.dart are analyzed.
     assertNotificationsText(r'''
 AnalysisErrors
-  file: /home/test/pubspec.yaml
-  errors: empty
-AnalysisErrors
   file: /home/test/lib/a.dart
   errors: notEmpty
 AnalysisErrors
@@ -124,12 +121,8 @@ analyzer:
 AnalysisFlush
   /home/test/lib/a.dart
   /home/test/lib/b.dart
-  /home/test/pubspec.yaml
 AnalysisErrors
   file: /home/test/analysis_options.yaml
-  errors: empty
-AnalysisErrors
-  file: /home/test/pubspec.yaml
   errors: empty
 AnalysisErrors
   file: /home/test/lib/a.dart
@@ -149,9 +142,6 @@ AnalysisErrors
     // a.dart was analyzed
     assertNotificationsText(r'''
 AnalysisErrors
-  file: /home/test/pubspec.yaml
-  errors: empty
-AnalysisErrors
   file: /home/test/lib/a.dart
   errors: notEmpty
 ''');
@@ -168,13 +158,9 @@ analyzer:
     assertNotificationsText(r'''
 AnalysisFlush
   /home/test/lib/a.dart
-  /home/test/pubspec.yaml
 AnalysisErrors
   file: /home/test/analysis_options.yaml
   errors: notEmpty
-AnalysisErrors
-  file: /home/test/pubspec.yaml
-  errors: empty
 AnalysisErrors
   file: /home/test/lib/a.dart
   errors: notEmpty
@@ -195,9 +181,6 @@ analyzer:
     assertNotificationsText(r'''
 AnalysisErrors
   file: /home/test/analysis_options.yaml
-  errors: empty
-AnalysisErrors
-  file: /home/test/pubspec.yaml
   errors: empty
 AnalysisErrors
   file: /home/test/lib/a.dart
@@ -232,9 +215,6 @@ AnalysisErrors
 AnalysisErrors
   file: /home/test/analysis_options.yaml
   errors: empty
-AnalysisErrors
-  file: /home/test/pubspec.yaml
-  errors: empty
 ''');
 
     _createFilesWithErrors([
@@ -267,9 +247,6 @@ void f(A a) {}
     assertNotificationsText(r'''
 AnalysisErrors
   file: /home/test/analysis_options.yaml
-  errors: empty
-AnalysisErrors
-  file: /home/test/pubspec.yaml
   errors: empty
 AnalysisErrors
   file: /home/test/lib/b.dart
@@ -315,9 +292,6 @@ AnalysisErrors
   file: /home/test/analysis_options.yaml
   errors: empty
 AnalysisErrors
-  file: /home/test/pubspec.yaml
-  errors: empty
-AnalysisErrors
   file: /home/test/lib/b.dart
   errors: notEmpty
 ''');
@@ -349,9 +323,6 @@ AnalysisErrors
     assertNotificationsText(r'''
 AnalysisErrors
   file: /home/test/analysis_options.yaml
-  errors: empty
-AnalysisErrors
-  file: /home/test/pubspec.yaml
   errors: empty
 AnalysisErrors
   file: /home/test/lib/a.dart
@@ -388,9 +359,6 @@ AnalysisErrors
     assertNotificationsText(r'''
 AnalysisErrors
   file: /home/test/analysis_options.yaml
-  errors: empty
-AnalysisErrors
-  file: /home/test/pubspec.yaml
   errors: empty
 AnalysisErrors
   file: /home/test/lib/a.dart
@@ -437,9 +405,6 @@ AnalysisErrors
   file: /home/test/analysis_options.yaml
   errors: empty
 AnalysisErrors
-  file: /home/test/pubspec.yaml
-  errors: empty
-AnalysisErrors
   file: /home/test/lib/test.dart
   errors: notEmpty
 ''');
@@ -460,12 +425,8 @@ AnalysisErrors
 AnalysisFlush
   /home/test/analysis_options.yaml
   /home/test/lib/test.dart
-  /home/test/pubspec.yaml
 AnalysisErrors
   file: /home/test/analysis_options.yaml
-  errors: empty
-AnalysisErrors
-  file: /home/test/pubspec.yaml
   errors: empty
 AnalysisErrors
   file: /home/test/lib/test.dart
@@ -492,9 +453,6 @@ AnalysisErrors
   file: /home/test/analysis_options.yaml
   errors: empty
 AnalysisErrors
-  file: /home/test/pubspec.yaml
-  errors: empty
-AnalysisErrors
   file: /home/test/lib/a.dart
   errors: notEmpty
 ''');
@@ -513,7 +471,6 @@ dependencies: true
 AnalysisFlush
   /home/test/analysis_options.yaml
   /home/test/lib/a.dart
-  /home/test/pubspec.yaml
 AnalysisErrors
   file: /home/test/analysis_options.yaml
   errors: empty
@@ -539,9 +496,6 @@ AnalysisErrors
     assertNotificationsText(r'''
 AnalysisErrors
   file: /home/test/analysis_options.yaml
-  errors: empty
-AnalysisErrors
-  file: /home/test/pubspec.yaml
   errors: empty
 AnalysisErrors
   file: /home/test/lib/a.dart
@@ -582,9 +536,6 @@ AnalysisErrors
   file: /home/test/analysis_options.yaml
   errors: empty
 AnalysisErrors
-  file: /home/test/pubspec.yaml
-  errors: empty
-AnalysisErrors
   file: /home/test/lib/a.dart
   errors: notEmpty
 AnalysisErrors
@@ -609,12 +560,8 @@ AnalysisFlush
   /home/test/analysis_options.yaml
   /home/test/lib/a.dart
   /home/test/lib/c.dart
-  /home/test/pubspec.yaml
 AnalysisErrors
   file: /home/test/analysis_options.yaml
-  errors: empty
-AnalysisErrors
-  file: /home/test/pubspec.yaml
   errors: empty
 AnalysisErrors
   file: /home/test/lib/a.dart
@@ -649,9 +596,6 @@ AnalysisErrors
 AnalysisErrors
   file: /home/test/AndroidManifest.xml
   errors: notEmpty
-AnalysisErrors
-  file: /home/test/pubspec.yaml
-  errors: empty
 AnalysisErrors
   file: /home/test/lib/a.dart
   errors: empty
@@ -693,9 +637,6 @@ void f(A a) {}
     assertNotificationsText(r'''
 AnalysisErrors
   file: /home/test/analysis_options.yaml
-  errors: empty
-AnalysisErrors
-  file: /home/test/pubspec.yaml
   errors: empty
 AnalysisErrors
   file: /home/test/lib/a.dart
@@ -743,9 +684,6 @@ void f(A a) {}
     assertNotificationsText(r'''
 AnalysisErrors
   file: /home/test/analysis_options.yaml
-  errors: empty
-AnalysisErrors
-  file: /home/test/pubspec.yaml
   errors: empty
 AnalysisErrors
   file: /home/test/lib/b.dart
@@ -797,9 +735,6 @@ AnalysisErrors
   file: /home/test/analysis_options.yaml
   errors: empty
 AnalysisErrors
-  file: /home/test/pubspec.yaml
-  errors: empty
-AnalysisErrors
   file: /home/test/lib/b.dart
   errors: notEmpty
 ''');
@@ -837,9 +772,6 @@ AnalysisErrors
 AnalysisErrors
   file: /home/test/lib/fix_data/foo.yaml
   errors: notEmpty
-AnalysisErrors
-  file: /home/test/pubspec.yaml
-  errors: empty
 AnalysisErrors
   file: /home/test/lib/a.dart
   errors: empty
@@ -886,9 +818,6 @@ AnalysisErrors
   file: /home/test/lib/fix_data.yaml
   errors: notEmpty
 AnalysisErrors
-  file: /home/test/pubspec.yaml
-  errors: empty
-AnalysisErrors
   file: /home/test/lib/a.dart
   errors: empty
 ''');
@@ -931,9 +860,6 @@ AnalysisErrors
     assertNotificationsText(r'''
 AnalysisErrors
   file: /home/test/analysis_options.yaml
-  errors: empty
-AnalysisErrors
-  file: /home/test/pubspec.yaml
   errors: empty
 AnalysisErrors
   file: /home/test/lib/test.dart
@@ -987,9 +913,6 @@ AnalysisErrors
   file: /home/test/analysis_options.yaml
   errors: empty
 AnalysisErrors
-  file: /home/test/pubspec.yaml
-  errors: empty
-AnalysisErrors
   file: /home/test/lib/test.dart
   errors: notEmpty
 ''');
@@ -1009,12 +932,8 @@ AnalysisErrors
 AnalysisFlush
   /home/test/analysis_options.yaml
   /home/test/lib/test.dart
-  /home/test/pubspec.yaml
 AnalysisErrors
   file: /home/test/analysis_options.yaml
-  errors: empty
-AnalysisErrors
-  file: /home/test/pubspec.yaml
   errors: empty
 AnalysisErrors
   file: /home/test/lib/test.dart
@@ -1045,9 +964,6 @@ AnalysisErrors
   file: /home/test/analysis_options.yaml
   errors: empty
 AnalysisErrors
-  file: /home/test/pubspec.yaml
-  errors: empty
-AnalysisErrors
   file: /home/test/lib/a.dart
   errors: notEmpty
 ''');
@@ -1063,10 +979,6 @@ AnalysisErrors
 AnalysisFlush
   /home/test/analysis_options.yaml
   /home/test/lib/a.dart
-  /home/test/pubspec.yaml
-AnalysisErrors
-  file: /home/test/pubspec.yaml
-  errors: empty
 AnalysisErrors
   file: /home/test/lib/a.dart
   errors: notEmpty
@@ -1102,9 +1014,6 @@ AnalysisErrors
   file: /home/test/AndroidManifest.xml
   errors: notEmpty
 AnalysisErrors
-  file: /home/test/pubspec.yaml
-  errors: empty
-AnalysisErrors
   file: /home/test/lib/a.dart
   errors: empty
 ''');
@@ -1137,9 +1046,6 @@ AnalysisFlush
     assertNotificationsText(r'''
 AnalysisErrors
   file: /home/test/analysis_options.yaml
-  errors: empty
-AnalysisErrors
-  file: /home/test/pubspec.yaml
   errors: empty
 AnalysisErrors
   file: /home/test/lib/a.dart
@@ -1213,9 +1119,6 @@ AnalysisErrors
   file: /home/test/analysis_options.yaml
   errors: empty
 AnalysisErrors
-  file: /home/test/pubspec.yaml
-  errors: empty
-AnalysisErrors
   file: /home/test/lib/a.dart
   errors: notEmpty
 ''');
@@ -1261,9 +1164,6 @@ AnalysisErrors
   file: /home/test/analysis_options.yaml
   errors: empty
 AnalysisErrors
-  file: /home/test/pubspec.yaml
-  errors: empty
-AnalysisErrors
   file: /home/test/lib/b.dart
   errors: empty
 ''');
@@ -1300,9 +1200,6 @@ AnalysisErrors
 AnalysisErrors
   file: /home/test/lib/fix_data.yaml
   errors: notEmpty
-AnalysisErrors
-  file: /home/test/pubspec.yaml
-  errors: empty
 AnalysisErrors
   file: /home/test/lib/a.dart
   errors: empty
@@ -1347,9 +1244,6 @@ AnalysisErrors
   file: /home/test/analysis_options.yaml
   errors: empty
 AnalysisErrors
-  file: /home/test/pubspec.yaml
-  errors: empty
-AnalysisErrors
   file: /home/test/lib/test.dart
   errors: empty
 ''');
@@ -1368,12 +1262,8 @@ AnalysisErrors
 AnalysisFlush
   /home/test/analysis_options.yaml
   /home/test/lib/test.dart
-  /home/test/pubspec.yaml
 AnalysisErrors
   file: /home/test/analysis_options.yaml
-  errors: empty
-AnalysisErrors
-  file: /home/test/pubspec.yaml
   errors: empty
 AnalysisErrors
   file: /home/test/lib/test.dart
@@ -1397,9 +1287,6 @@ AnalysisErrors
     assertNotificationsText(r'''
 AnalysisErrors
   file: /home/test/analysis_options.yaml
-  errors: empty
-AnalysisErrors
-  file: /home/test/pubspec.yaml
   errors: empty
 AnalysisErrors
   file: /home/test/lib/a.dart
@@ -1468,9 +1355,6 @@ AnalysisErrors
     assertNotificationsText(r'''
 AnalysisErrors
   file: /home/test/analysis_options.yaml
-  errors: empty
-AnalysisErrors
-  file: /home/test/pubspec.yaml
   errors: empty
 AnalysisErrors
   file: /home/foo/lib/a.dart
@@ -1629,9 +1513,6 @@ AnalysisErrors
   file: /home/test/analysis_options.yaml
   errors: empty
 AnalysisErrors
-  file: /home/test/pubspec.yaml
-  errors: empty
-AnalysisErrors
   file: /home/test/lib/a.dart
   errors: notEmpty
 AnalysisErrors
@@ -1693,9 +1574,6 @@ AnalysisErrors
   file: /home/test/analysis_options.yaml
   errors: empty
 AnalysisErrors
-  file: /home/test/pubspec.yaml
-  errors: empty
-AnalysisErrors
   file: /home/test/lib/a.dart
   errors: notEmpty
 ''');
@@ -1733,9 +1611,6 @@ AnalysisErrors
     assertNotificationsText(r'''
 AnalysisErrors
   file: /home/test/analysis_options.yaml
-  errors: empty
-AnalysisErrors
-  file: /home/test/pubspec.yaml
   errors: empty
 AnalysisErrors
   file: /home/test/lib/a.dart
@@ -1786,9 +1661,6 @@ analyzer:
 
     assertNotificationsText(r'''
 AnalysisErrors
-  file: /home/test/pubspec.yaml
-  errors: empty
-AnalysisErrors
   file: /home/test/lib/a.dart
   errors: notEmpty
 ''');
@@ -1817,9 +1689,6 @@ AnalysisErrors
 AnalysisErrors
   file: /home/test/AndroidManifest.xml
   errors: notEmpty
-AnalysisErrors
-  file: /home/test/pubspec.yaml
-  errors: empty
 ''');
   }
 
@@ -1838,9 +1707,6 @@ AnalysisErrors
 AnalysisErrors
   file: /home/test/lib/fix_data.yaml
   errors: notEmpty
-AnalysisErrors
-  file: /home/test/pubspec.yaml
-  errors: empty
 ''');
   }
 
@@ -1857,9 +1723,6 @@ AnalysisErrors
 AnalysisErrors
   file: /home/test/lib/fix_data.yaml
   errors: empty
-AnalysisErrors
-  file: /home/test/pubspec.yaml
-  errors: empty
 ''');
   }
 
@@ -1875,9 +1738,6 @@ AnalysisErrors
   errors: empty
 AnalysisErrors
   file: /home/test/lib/fix_data.yaml
-  errors: empty
-AnalysisErrors
-  file: /home/test/pubspec.yaml
   errors: empty
 ''');
   }
@@ -1944,9 +1804,6 @@ AnalysisErrors
   file: /home/test/analysis_options.yaml
   errors: empty
 AnalysisErrors
-  file: /home/test/pubspec.yaml
-  errors: empty
-AnalysisErrors
   file: /home/test/lib/test.dart
   errors: empty
 ''');
@@ -1962,9 +1819,6 @@ AnalysisErrors
     assertNotificationsText(r'''
 AnalysisErrors
   file: /home/test/analysis_options.yaml
-  errors: empty
-AnalysisErrors
-  file: /home/test/pubspec.yaml
   errors: empty
 AnalysisErrors
   file: /home/test/lib/test.dart
@@ -2004,9 +1858,6 @@ AnalysisErrors
     assertNotificationsText(r'''
 AnalysisErrors
   file: /home/test/analysis_options.yaml
-  errors: empty
-AnalysisErrors
-  file: /home/test/pubspec.yaml
   errors: empty
 AnalysisErrors
   file: /home/test/lib/test.dart